Add CRUD Users

+++ app/Http/Controllers/UserController.php
@@ -12,6 +12,12 @@ class UserController extends Controller
         return view('users.index');
     }
 
+    public function show()
+    {
+        $users = User::all();
+        return response()->json(['data' => $users]);
+    }
+
     public function create()
     {
         return view('users.create');
@@ -19,25 +25,52 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
-        $user = User::create($request->all());
-        return response()->json($user);
+        if($request->ajax()) {
+            $request->validate([
+                'name' => 'required|string|max:255',
+                'email' => 'required|string|email|max:255|unique:users',
+                'password' => 'required|string|min:8',
+                'password_confirmation' => 'required|string|min:8|same:password',
+            ]);
+
+            User::create([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password)
+            ]);
+            
+            return response()->json([
+                'success' => true,
+                'message' => 'User data has been successfully saved!',
+                'type' => 'success'
+            ]);
+        }
+    }
+
+    public function edit(User $user)
+    {
+        return view('users.edit', compact('user'));
     }
 
     public function update(Request $request, User $user)
     {
-        $user->update($request->all());
-        return response()->json($user);
+        if($request->ajax()) {
+            $user->update($request->all());
+            return response()->json($user);
+        }
     }
 
-    public function destroy(User $user)
+    public function destroy(Request $request, User $user)
     {
+        if($request->ajax()) {
         $user->delete();
-        return response()->json(['success' => true]);
+            return response()->json([
+                'success' => true,
+                'message' => 'User data has been successfully deleted!',
+                'type' => 'success'
+            ]);
+        }
     }
 
-    public function show()
-    {
-        $users = User::all();
-        return response()->json(['data' => $users]);
-    }
+    
 }
